+++ apps/Luxelot.Apps.FserveApp/ClientConnection.cs
@@ -3,22 +3,13 @@ using System.Text;
 
 namespace Luxelot.Apps.FserveApp;
 
-public class ClientConnection
+public class ClientConnection(ImmutableArray<byte> sessionSharedKey)
 {
-    public ImmutableArray<byte> SessionSharedKey { get; private set; }
+    public ImmutableArray<byte> SessionSharedKey { get; private set; } = sessionSharedKey;
 
     public ImmutableArray<byte>? Principal { get; set; }
 
     public string? PrincipalAsName { get => Principal == null ? null : Encoding.UTF8.GetString([.. Principal]); }
 
     public string CurrentWorkingDirectory { get; set; } = "/";
-
-    public ClientConnection(ImmutableArray<byte> sessionSharedKey)
-    {
-        ArgumentNullException.ThrowIfNull(sessionSharedKey);
-
-        SessionSharedKey = sessionSharedKey;
-    }
-
-
 }