+++ apps/Luxelot.Apps.PingApp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.2 - 2024-11-26
+
+### Changed
+
+- Upgraded to .NET 9.0 SDK
+
+
 ## 1.1 - 2024-11-24
 
 ### Changed